organize data folder and update imports
