+++ src/components/FeaturedHerbTeaser.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { motion } from 'framer-motion'
 import { Link } from 'react-router-dom'
-import { herbs } from '../../herbsfull'
+import { herbs } from '../data/herbsfull'
 import type { Herb } from '../types'
 import TagBadge from './TagBadge'
 import { slugify } from '../utils/slugify'
