+++ src/data/blurbs.ts
