+++ src/data/index.ts
@@ -1,0 +1,4 @@
+export { default as herbsData } from "./herbs.json";
+export { herbs } from "./herbsfull";
+export { herbBlurbs } from "./blurbs";
+export { compounds } from "./compounds";
