+++ src/pages/Downloads.tsx
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet-async'
 import { saveAs } from 'file-saver'
 import jsPDF from 'jspdf'
 import { Herb } from '../types'
-import { herbs } from '../../herbsfull'
+import { herbs } from '../data/herbsfull'
 
 const allHerbs = herbs
