+++ src/pages/HerbDetail.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useParams, Link } from 'react-router-dom'
 import { motion } from 'framer-motion'
 import { Helmet } from 'react-helmet-async'
-import { herbs } from '../../herbsfull'
+import { herbs } from '../data/herbsfull'
 import { decodeTag, tagVariant } from '../utils/format'
 import TagBadge from '../components/TagBadge'
 import CompoundBadge from '../components/CompoundBadge'
