+++ src/pages/HerbDetailView.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useParams, Link } from 'react-router-dom'
 import { motion } from 'framer-motion'
 import { Helmet } from 'react-helmet-async'
-import { herbs } from '../../herbsfull'
+import { herbs } from '../data/herbsfull'
 import TagBadge from '../components/TagBadge'
 import { decodeTag, tagVariant } from '../utils/format'
 import { slugify } from '../utils/slugify'
